Rename merkle module
