+++ crates/kilt-dip-primitives/src/lib.rs
@@ -27,7 +27,7 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 /// Module to deal with cross-chain Merkle proof as generated by the KILT chain.
-pub mod merkle;
+pub mod merkle_proofs;
 /// Module to deal with cross-chain state proofs.
 pub mod state_proofs;
 /// Collection of traits used throughout the crate and useful for both providers
@@ -38,6 +38,6 @@ pub mod utils;
 /// deployed both on a sibling parachain and on a parent relaychain.
 pub mod verifier;
 
-pub use merkle::latest::*;
+pub use merkle_proofs::latest::*;
 pub use traits::RelayStateRootsViaRelayStorePallet;
 pub use verifier::*;
