+++ crates/kilt-dip-primitives/src/merkle_proofs/v0/error.rs
