+++ crates/kilt-dip-primitives/src/verifier/parachain/mod.rs
@@ -27,7 +27,7 @@ use scale_info::TypeInfo;
 use sp_std::{fmt::Debug, marker::PhantomData, vec::Vec};
 
 use crate::{
-	merkle::v0::RevealedDidKey,
+	merkle_proofs::v0::RevealedDidKey,
 	traits::{DipCallOriginFilter, GetWithArg, GetWithoutArg, Incrementable},
 	utils::OutputOf,
 	DipOriginInfo, ParachainDipDidProof,
